LayoutTests/media/video-play-empty-events.html

<video controls></video>
<p>Test that play() from EMPTY network state triggers load() and async play event.</p>
<script src=video-test.js></script>
<script>
video.src = "content/test.mp4";
test("video.networkState == HTMLMediaElement.EMPTY");
waitForEventAndTest('loadstart', "video.paused");
waitForEventTestAndEnd('play', "!video.paused");
run("video.play()");
consoleWrite("SCRIPT DONE");
</script>
