LayoutTests/media/video-play-empty-events-expected.txt

Test that play() from EMPTY network state triggers load() and async play event.

TEST(video.networkState == HTMLMediaElement.EMPTY) OK
RUN(video.play())
EVENT(loadstart) TEST(video.paused) OK
SCRIPT DONE
EVENT(play) TEST(!video.paused) OK
END OF TEST
